# backend/docs/swagger.yaml
openapi: 3.0.0
servers:
    - description: localhost:5000
      url: http://localhost:5000/api
info:
    description: Plant care system
    version: '1.0.0'
    title: Plan-ts API
    contact:
        email: 'imgjeits@stud.ntnu.no'
    license:
        name: Attribution-NoDerivs 3.0 Unported (CC BY-ND 3.0)
        url: 'https://creativecommons.org/licenses/by-nd/3.0/'
...
paths:
    /plants:
        get:
            summary: Lists plants
            tags:
                - all
                - plants
            description: Get a list of plants in the system, sorted alphabetically by name. List can be filtered by using queries. No query will return a list of all the plants.
            operationId: listPlants
            responses:
                '200':
                    description: List of plants matching criteria
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    properties:
                                        health:
                                            $ref: '#/components/schemas/Plant/properties/health'
                                        picture:
                                            $ref: '#/components/schemas/Plant/properties/picture'
                                        _id:
                                            type: string
                                            example: '607dde78081f316f0473f7fb'
                                        name:
                                            $ref: '#/components/schemas/Plant/properties/name'
                                        location:
                                            $ref: '#/components/schemas/Location'
                                        createdAt:
                                            $ref: '#/components/schemas/Plant/properties/createdAt'
                                        updatedAt:
                                            $ref: '#/components/schemas/Plant/properties/updatedAt'
                                        __v:
                                            type: number
                '404':
                    description: Could not find any plants based on query
                '500':
                    description: Something went wrong on the server. Will include error message from server
        post:
            summary: Add new plant
            tags:
                - managers
                - plants
            description: Create a new plant and add it to the database.
            operationId: createPlant
            security:
                - bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    $ref: '#/components/schemas/Plant/properties/name'
                                location:
                                    $ref: '#/components/schemas/Plant/properties/location'
                                picture:
                                    $ref: '#/components/schemas/Plant/properties/picture'
                                notes:
                                    $ref: '#/components/schemas/Plant/properties/notes'
                                waterDaysBetween:
                                    $ref: '#/components/schemas/Plant/properties/health/properties/water/properties/days_between'
                                waterAmount:
                                    $ref: '#/components/schemas/Plant/properties/health/properties/water/properties/amount'
                                waterDue:
                                    $ref: '#/components/schemas/Plant/properties/health/properties/water/properties/due'
                                fertilizerDaysBetween:
                                    $ref: '#/components/schemas/Plant/properties/health/properties/fertilizer/properties/days_between'
                                fertilizerAmount:
                                    $ref: '#/components/schemas/Plant/properties/health/properties/fertilizer/properties/days_between'
                                fertilizerDue:
                                    $ref: '#/components/schemas/Plant/properties/health/properties/fertilizer/properties/days_between'
                                lightAmount:
                                    $ref: '#/components/schemas/Plant/properties/health/properties/light/properties/amount'
            responses:
                '201':
                    description: Successful plant creation
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                        example: 'New plant successfully created.'
                                    plant:
                                        properties:
                                            _id:
                                                type: string
                                            name:
                                                type: string
                                            location:
                                                type: string
                                            picture:
                                                type: string
                                            health:
                                                type: object
                                                properties:
                                                    water:
                                                        type: object
                                                        properties:
                                                            days_between:
                                                                type: number
                                                            amount:
                                                                type: number
                                                            due:
                                                                type: string
                                                                format: date-time
                                                    fertilizer:
                                                        type: object
                                                        properties:
                                                            days_between:
                                                                type: number
                                                            amount:
                                                                type: number
                                                            due:
                                                                type: string
                                                                format: date-time
                                                    light:
                                                        type: object
                                                        properties:
                                                            amount:
                                                                type: number
                                            createdAt:
                                                type: string
                                                format: date-time
                                            updatedAt:
                                                type: string
                                                format: date-time
                                            __v:
                                                type: number
                '401':
                    $ref: '#/components/responses/UnauthorizedError'
                '500':
                    description: Something went wrong on the server. Will include error message from server
    /plants/notifications:
        get:
            summary: Get plants that have tasks due
            tags:
                - gardeners
                - plants
            description: Get array of plant objects where the plant needs watering/fertilizing. Will return one array for each of the types. Can also find plants that have tasks due between now and the next *x* amount of days.
            operationId: pastDue
            security:
                - bearerAuth: []
            parameters:
                - in: query
                  name: days
                  description: How many days ahead
                  required: false
                  schema:
                      type: string
                      example: 2
            responses:
                '200':
                    description: Successful query returning two arrays
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    water:
                                        type: array
                                        items:
                                            type: object
                                            properties:
                                                health:
                                                    type: object
                                                    properties:
                                                        water:
                                                            $ref: '#/components/schemas/Plant/properties/health/properties/water'
                                                _id:
                                                    type: string
                                                    example: '6075b45e72d09a3aacc2eafc'
                                                name:
                                                    $ref: '#/components/schemas/Plant/properties/name'
                                                location:
                                                    $ref: '#/components/schemas/Location'
                                    fertilize:
                                        type: array
                                        items:
                                            type: object
                                            properties:
                                                health:
                                                    type: object
                                                    properties:
                                                        water:
                                                            $ref: '#/components/schemas/Plant/properties/health/properties/water'
                                                _id:
                                                    type: string
                                                    example: '6075b45e72d09a3aacc2eafc'
                                                name:
                                                    $ref: '#/components/schemas/Plant/properties/name'
                                                location:
                                                    $ref: '#/components/schemas/Location'
                '401':
                    $ref: '#/components/responses/UnauthorizedError'
                '404':
                    description: No plants with tasks due
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: 'Could not find any plants with tasks past due at this time.'
                '500':
                    description: Server error
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: 'Something went wrong while trying to find plants. [SERVER ERROR MESSAGE]'
    /plants/notifications/count:
        get:
            summary: Get number of notifications
            tags:
                - gardeners
                - plants
            description: Counts the number of tasks due (not the number of plants)
            operationId: noOfNotifications
            security:
                - bearerAuth: []
            responses:
                '200':
                    description: Successful query returning object with count
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    count:
                                        type: number
                                        example: 3
                '401':
                    $ref: '#/components/responses/UnauthorizedError'
                '404':
                    description: No tasks due
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: 'Could not find any tasks due at this time.'
                '500':
                    description: Server error
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: 'Something went wrong while trying to find tasks due. [SERVER ERROR MESSAGE]'
    /plants/search:
        get:
            summary: Search plants by name
            tags:
                - all
                - plants
            description: Search for plant by name, or parts of the name
            operationId: searchPlants
            parameters:
                - in: query
                  name: searchField
                  description: Search string
                  required: true
                  schema:
                      type: string
                      example: 'succulent'
            responses:
                '200':
                    description: Successful search
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    properties:
                                        health:
                                            $ref: '#/components/schemas/Plant/properties/health'
                                        picture:
                                            $ref: '#/components/schemas/Plant/properties/picture'
                                        _id:
                                            type: string
                                            example: '607dde78081f316f0473f7fb'
                                        name:
                                            $ref: '#/components/schemas/Plant/properties/name'
                                        location:
                                            $ref: '#/components/schemas/Location'
                                        createdAt:
                                            $ref: '#/components/schemas/Plant/properties/createdAt'
                                        updatedAt:
                                            $ref: '#/components/schemas/Plant/properties/updatedAt'
                                        __v:
                                            type: number
                '400':
                    description: No search string
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: 'Query needs to include a searchField value.'
                '404':
                    description: No plants found
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: 'Search did not match any of the plants.'
                '500':
                    description: Server error
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: 'Something went wrong while trying to search for plants. [SERVER ERROR MESSAGE].'
...
    /plants/{id}/notes:
        get:
            summary: Get plant notes
            tags:
                - gardeners
                - plants
            description: Get notes connected to specific plant by its ID
            operationId: plantNotes
            security:
                - bearerAuth: []
            parameters:
                - in: path
                  name: id
                  description: ObjectID of the plant
                  required: true
                  schema:
                      type: string
                      example: '6086c0b2e00c2a888c26af05'
            responses:
                '200':
                    description: Successful query returning notes object
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    notes:
                                        $ref: '#/components/schemas/Plant/properties/notes'
                '401':
                    $ref: '#/components/responses/UnauthorizedError'
                '404':
                    description: Could not find the plant
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: 'Plant not found.'
                '500':
                    description: Server error
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                                        example: 'Something went wrong while looking for notes connected to plant with ID 607dde78081f316f0473f7fb. [SERVER ERROR MESSAGE].'
...
components:
...
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    responses:
        UnauthorizedError:
            description: Access token is missing or invalid